+++ src/app.css
@@ -35,7 +35,7 @@
 }
 
 .userButtons:hover {
-  background: rgb(116, 116, 116);
+  background: #3f3f3f;
   color: white;
 }
