+++ src/components/Users/Users.css
@@ -1,0 +1,24 @@
+.userCards {
+  display: flex;
+  flex-direction: column;
+}
+
+.userCards * {
+  padding: 10px;
+}
+
+.user {
+  border: 1px solid #eee;
+}
+
+.user button {
+  background-color: #f7f7f7;
+  font-size: 20px;
+  border: 3px solid black;
+  border-radius: 10px;
+}
+
+.user button:hover {
+  background-color: #3f3f3f;
+  color: white;
+}
